Support TYPO3v10, drop support for TYPO3v9
1. Middleware order
Biggest problem with v10 is that the order of the middlewares changed.
In TYPO3v9, "typo3/cms-frontend/authentication" came first, and
"typo3/cms-frontend/backend-user-authentication" second.
simulatebe could squeeze between them.

With TYPO3v10 the order is switched:
1. typo3/cms-frontend/backend-user-authentication
2. typo3/cms-frontend/authentication

This means we don't have a frontend user when we are called before
backend-user-authentication, and can't rely on backend-user-authentication
to initialize the backend user automatically after we simulated
the backend user cookie.

Instead we call the backend-user-authentication middleware ourselves
to let it handle the setup, even if it's called a second time now.

2. Logout process
We have to rely on the logoff_post_processing hook to logout the backend user
when the frontend user is logged out, because we do not know that
a frontend user existed when our middleware is called.

Cookie cleanup is still done in our middleware because that's easier
to do than in the logoff hook.

Resolves: https://github.com/CDRO/simulatebe/issues/15